+++ src/modules/ngx-select/component/index.ts
@@ -1,4 +1,4 @@
-export * from './ngx-select';
+export * from './ngx-select.component';
 export * from './ngx-select-template';
-export * from './ngx-select-option/ngx-select-option';
+export * from './ngx-select-option/ngx-select-option.component';
 export * from './ngx-select-model';
