+++ src/modules/ngx-select/component/ngx-select-option/ngx-select-option.component.html
