+++ src/modules/ngx-select/component/ngx-select.component.css
