+++ src/modules/ngx-select/ngx-select.module.ts
@@ -2,8 +2,8 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { OverlayModule } from '@angular/cdk/overlay';
 
-import { NgxSelectComponent } from './component/ngx-select';
-import { NgxSelectOptionComponent } from './component/ngx-select-option/ngx-select-option';
+import { NgxSelectComponent } from './component/ngx-select.component';
+import { NgxSelectOptionComponent } from './component/ngx-select-option/ngx-select-option.component';
 import { NgxSelectTemplate } from './component/ngx-select-template';
 
 @NgModule({
